Added abstract class for creating logically grouped api test cases
