+++ tests/Jira/ApiTest.php
@@ -4,55 +4,16 @@ namespace Tests\chobie\Jira;
 
 
 use chobie\Jira\Api;
-use chobie\Jira\Api\Authentication\AuthenticationInterface;
-use chobie\Jira\Api\Exception;
 use chobie\Jira\Api\Result;
 use chobie\Jira\IssueType;
-use Prophecy\Prophecy\ObjectProphecy;
-use chobie\Jira\Api\Client\ClientInterface;
 
 /**
  * Class ApiTest
  *
  * @package Tests\chobie\Jira
  */
-class ApiTest extends AbstractTestCase
+class ApiTest extends AbstractApiTest
 {
-
-	const ENDPOINT = 'http://jira.company.com';
-
-	/**
-	 * Api.
-	 *
-	 * @var Api
-	 */
-	protected $api;
-
-	/**
-	 * Credential.
-	 *
-	 * @var AuthenticationInterface
-	 */
-	protected $credential;
-
-	/**
-	 * Client.
-	 *
-	 * @var ObjectProphecy
-	 */
-	protected $client;
-
-	/**
-	 * @before
-	 */
-	protected function setUpTest()
-	{
-		$this->credential = $this->prophesize(AuthenticationInterface::class)->reveal();
-		$this->client = $this->prophesize(ClientInterface::class);
-
-		$this->api = new Api(self::ENDPOINT, $this->credential, $this->client->reveal());
-		$this->api->setOptions(0); // Disable automapping.
-	}
 
 	/**
 	 * @dataProvider setEndpointDataProvider
@@ -784,50 +745,6 @@ class ApiTest extends AbstractTestCase
 		$actual = $this->api->deleteWorklog('JRA-15', 11256, array('custom' => 'param'));
 
 		$this->assertEquals(json_decode($response, true), $actual, 'The response is json-decoded.');
-	}
-
-	/**
-	 * Checks, that response is correct.
-	 *
-	 * @param string       $expected_raw_response Expected raw response.
-	 * @param Result|false $actual_response       Actual response.
-	 *
-	 * @return void
-	 */
-	protected function assertApiResponse($expected_raw_response, $actual_response)
-	{
-		$expected = new Result(json_decode($expected_raw_response, true));
-
-		// You'll get "false", when unexpected API call was made.
-		if ( $actual_response !== false ) {
-			$this->assertEquals($expected, $actual_response);
-		}
-	}
-
-	/**
-	 * Expects a particular client call.
-	 *
-	 * @param string       $method       Request method.
-	 * @param string       $url          URL.
-	 * @param array|string $data         Request data.
-	 * @param string       $return_value Return value.
-	 * @param boolean      $is_file      This is a file upload request.
-	 * @param boolean      $debug        Debug this request.
-	 *
-	 * @return void
-	 */
-	protected function expectClientCall(
-		$method,
-		$url,
-		$data = array(),
-		$return_value,
-		$is_file = false,
-		$debug = false
-	) {
-		$this->client
-			->sendRequest($method, $url, $data, self::ENDPOINT, $this->credential, $is_file, $debug)
-			->willReturn($return_value)
-			->shouldBeCalled();
 	}
 
 }
